[BUGFIX] Keep category with same name on project change
To make the behavior of Redmine consistent when editing the
project of a single issue the category of an issue will not be
removed when a category with the same name is found in the
new project.

When editing multiple issues Redmine already keeps the category
with the matching name.

Resolves: #16941